Move TxUrl assembly into a separate module
Creating transactions' URLs is a functionality that doesn't need to be
tightly-coupled with other modules.
